requirement.txt
amqp==2.5.2
billiard==3.6.2.0
boto3==1.10.30
botocore==1.13.50
celery==4.4.0
certifi==2019.11.28
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
Django==2.0
django-cors-headers==3.2.0
djangorestframework==3.10.3
docutils==0.15.2
drf-yasg==1.17.0
idna==2.8
importlib-metadata==1.5.0
inflection==0.3.1
itypes==1.1.0
Jinja2==2.11.1
jmespath==0.9.4
kombu==4.6.7
MarkupSafe==1.1.1
packaging==20.1
Pillow==7.0.0
psycopg2==2.8.4
PyJWT==1.7.1
pyparsing==2.4.6
python-dateutil==2.8.1
python-decouple==3.3
pytz==2019.3
requests==2.22.0
ruamel.yaml==0.16.7
ruamel.yaml.clib==0.2.0
s3transfer==0.2.1
six==1.14.0
uritemplate==3.0.1
urllib3==1.25.8
vine==1.3.0
zipp==2.1.0
